remove env
